feat(download): add flag to avoid downloading existing images
### Description:
Every time we run the fad tool it seems that the existing images get overwritten
with something that makes git count them
as a new image.

### Solution:
Add a new flag (`-u`) and it's corresponding manifest flag
(`download_only_unexisting_in_folder`) to avoid overwritting the images
that we already have in the download folder.

Note that we don't check if there's a different version of each image,
we only check if the same image name exists.

Fixes: https://github.com/robertohuertasm/figma-asset-downloader/issues/18

CO-418